09._Environment_Variables/python/pyproject.toml
[tool.poetry]
name = "python_env"
version = "0.1.0"
description = ""
authors = ["rgodding <rashaun.godding@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
